+++ src/web/index.js
@@ -1,8 +1,10 @@
 
-export {default as LoginForm} from './components/LoginForm';
-export {default as SignUpForm} from './components/SignUpForm';
+export {default as auth} from './auth';
 export {default as ForgotPasswordForm} from './components/ForgotPasswordForm';
 export {default as ForgotPasswordRequestForm} from './components/ForgotPasswordRequestForm';
 export {default as ForgotPasswordConfirmForm} from './components/ForgotPasswordConfirmForm';
-export {default as auth} from './auth';
+export {default as LoginForm} from './components/LoginForm';
+export {default as SignUpForm} from './components/SignUpForm';
+export {default as SignUpRequestForm} from './components/SignUpRequestForm';
+export {default as VerificationForm} from './components/VerificationForm';
